+++ lib/sentry/integrations/oban/cron.ex
@@ -1,7 +1,23 @@
 defmodule Sentry.Integrations.Oban.Cron do
-  @moduledoc false
-  alias Sentry.CheckIn
+  @moduledoc """
+  This module provides built-in integration for cron jobs managed by Oban.
+  """
+
+  @moduledoc since: "10.9.0"
+
   alias Sentry.Integrations.CheckInIDMappings
+
+  @doc """
+  The Oban integration calls this callback (if present) to customize
+  the configuration options for the check-in.
+
+  This function must return options compatible with the ones passed to `Sentry.CheckIn.new/1`.
+
+  Options returned by this function overwrite any option inferred by the specific
+  integration for the check in. We perform *deep merging* of nested keyword options.
+  """
+  @doc since: "10.9.0"
+  @callback sentry_check_in_configuration(oban_job :: struct()) :: options_to_merge :: keyword()
 
   @events [
     [:oban, :job, :start],
@@ -9,12 +25,14 @@ defmodule Sentry.Integrations.Oban.Cron do
     [:oban, :job, :exception]
   ]
 
+  @doc false
   @spec attach_telemetry_handler(keyword()) :: :ok
   def attach_telemetry_handler(config) when is_list(config) do
     _ = :telemetry.attach_many(__MODULE__, @events, &__MODULE__.handle_event/4, config)
     :ok
   end
 
+  @doc false
   @spec handle_event([atom()], term(), term(), keyword()) :: :ok
   def handle_event(event, measurements, metadata, config)
 
@@ -107,7 +125,7 @@ defmodule Sentry.Integrations.Oban.Cron do
   else
     worker ->
       if Code.loaded?(worker) do
-        CheckIn.__resolve_custom_options__(opts, worker, job)
+        resolve_custom_opts(opts, worker, job)
       else
         opts
       end
@@ -115,6 +133,27 @@ defmodule Sentry.Integrations.Oban.Cron do
 
   defp resolve_custom_opts(opts, _job) do
     opts
+  end
+
+  defp resolve_custom_opts(options, mod, per_integration_term) do
+    custom_opts =
+      if function_exported?(mod, :sentry_check_in_configuration, 1) do
+        mod.sentry_check_in_configuration(per_integration_term)
+      else
+        []
+      end
+
+    deep_merge_keyword(options, custom_opts)
+  end
+
+  defp deep_merge_keyword(left, right) do
+    Keyword.merge(left, right, fn _key, left_val, right_val ->
+      if Keyword.keyword?(left_val) and Keyword.keyword?(right_val) do
+        deep_merge_keyword(left_val, right_val)
+      else
+        right_val
+      end
+    end)
   end
 
   defp schedule_opts(%{meta: meta} = job) when is_struct(job, Oban.Job) do
